# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel News API",
        "description": "API for aggregating news from multiple sources",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/v1/news-articles": {
            "get": {
                "tags": [
                    "News Articles"
                ],
                "summary": "Get all news articles",
                "operationId": "ab3941e6a7c72a5be92ffc3ee93f526a",
                "parameters": [
                    {
                        "name": "title[like]",
                        "in": "query",
                        "description": "Filter title using LIKE operator (use % for wildcards)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "%news%"
                        }
                    },
                    {
                        "name": "title[eq]",
                        "in": "query",
                        "description": "Filter title using exact match",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category[eq]",
                        "in": "query",
                        "description": "Filter by exact category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "general",
                                "business",
                                "technology",
                                "sports",
                                "science",
                                "health"
                            ]
                        }
                    },
                    {
                        "name": "source[eq]",
                        "in": "query",
                        "description": "Filter by exact source",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "NY Times"
                        }
                    },
                    {
                        "name": "author[like]",
                        "in": "query",
                        "description": "Filter author using LIKE operator",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "%John%"
                        }
                    },
                    {
                        "name": "published_at[gt]",
                        "in": "query",
                        "description": "Filter articles published after date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "published_at[lt]",
                        "in": "query",
                        "description": "Filter articles published before date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of news articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/NewsArticle"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/news-articles/{id}": {
            "get": {
                "tags": [
                    "News Articles"
                ],
                "summary": "Get a specific news article",
                "operationId": "1de777df5047743d2f542ce4f036e11e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "News article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "News article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsArticle"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "News article not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "2c518bfeba5f98da9877464becc61961",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "auth_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in to existing user",
                "operationId": "78efbe54c906001402450fe42c09b297",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "auth_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out the current user",
                "operationId": "025bceddd81a915fa73c021cdc4329ed",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/password/reset": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset user password",
                "operationId": "192bdd495cbb6e0e9223d84153226487",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset link sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/v1/preferences": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get user preferences",
                "operationId": "e633f9a25a1c59b17495086bb8da179d",
                "responses": {
                    "200": {
                        "description": "User preferences"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Update user preferences",
                "operationId": "20ddc2c0a51b79a94ba514bcde8fa21c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "sources": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "authors": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preferences updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/personalizedFeed": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get personalized news feed",
                "operationId": "154c462f96856a730dd378c7dc918b9e",
                "responses": {
                    "200": {
                        "description": "Personalized news feed"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "NewsArticle": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "News Title"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://example.com/news"
                    },
                    "source": {
                        "type": "string",
                        "example": "News Source"
                    },
                    "category": {
                        "type": "string",
                        "example": "general"
                    },
                    "author": {
                        "type": "string",
                        "example": "Author Name"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Endpoints for user authentication"
        },
        {
            "name": "News Articles",
            "description": "Endpoints for managing news articles"
        },
        {
            "name": "User Preferences",
            "description": "Endpoints for managing user preferences"
        }
    ]
}